healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: 3a62f8c3caf0ab097a2b0a9d2b291f15926a9245
Former-commit-id: 1856ba8bb623aa0502592acbd16904331df60511
Former-commit-id: 3317d630c2bfe55f03ef922d9bbdf0ab15d9ae6f