init: Weaken property override security for the init extension
Sometimes we need to override ro.* properties by using our vendor init
extension.

Previously there was a security check which was blocking that.
To resolve the issue, we need to weaken the security check during the
execution of our vendor init extension.

This is safe because the vendor init extension gets executed as part of init
construction and it is considered a trusted system component.

Change-Id: I6198b453745cb92c65d3e3d49e3262354cddd2a2
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Jayant-Deshmukh <jayantdeshmuk008@gmail.com>
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 15aafc0c179544dd173a7afdce8d4d24b19219e7
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: d5ce44bfdf41cbe51f7432a04409bb11147b35a4
Former-commit-id: e9481910a973d5dddf987f38250d78b92c0084ca
Former-commit-id: 37609514b2025ea5fc0314658109072780159e77
Former-commit-id: d53362e4fb95211f8eb77dd953aceed9972a7fec