libfs_avb: Disable dm-verity when AVB is permissive
When the bootloader is unlocked (i.e. AVB is permissive), enforcing
dm-verity on system partitions is meaningless because the bootloader
doesn't enforce verification on the root of the high-level verified boot
chain: the kernel. As a result, mounting system partitions with
dm-verity (hashtree verification) is futile when the code performing
verification has not been verified in the first place; users can also
disable dm-verity manually by flashing vbmeta with `fastboot flash
--disable-verity vbmeta vbmeta.img`.

For user and developer convenience, disable dm-verity automatically when
the bootloader is unlocked by checking for permissive AVB. This makes it
possible to ship enforcing vbmeta images for security-conscious users to
lock their bootloader and reap the benefits of verified boot, while
still allowing users with unlocked bootloaders to modify system
partitions.

Change-Id: I101903e17d06e07da29813cdb03973944629c760

Former-commit-id: 03e732282bd1b0ac2473e3b2c20fbc0b216ff112
Former-commit-id: 35f5a5e51ce5aae0adfaa910c1e99d97359be124
Former-commit-id: 7b708339bea5b543ad1ea5264ed346696fcfe06a
Former-commit-id: 3ff2cf8813e5b33d3768d4853525640e8bd07ad7